Refactor: Delivery subpackaging
footer: repository tests
