+++ src/main/java/shop/S5G/shop/controller/DeliveryController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import shop.S5G.shop.dto.delivery.DeliveryResponseDto;
 import shop.S5G.shop.dto.delivery.DeliveryUpdateRequestDto;
-import shop.S5G.shop.service.order.DeliveryService;
+import shop.S5G.shop.service.delivery.DeliveryService;
 
 @RequiredArgsConstructor
 @RequestMapping("/api/shop/delivery")
