+++ src/main/java/shop/S5G/shop/dto/delivery/DeliveryStatusView.java
@@ -1,0 +1,6 @@
+package shop.S5G.shop.dto.delivery;
+
+public interface DeliveryStatusView {
+    long getId();
+    String getName();
+}
