+++ src/main/java/shop/S5G/shop/entity/Book.java
@@ -1,7 +1,13 @@
 package shop.S5G.shop.entity;
 
-import jakarta.persistence.*;
-
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.AccessLevel;
@@ -38,6 +44,8 @@ public class Book {
     private BigDecimal discountRate;
     @Column(name = "is_packed")
     private boolean isPacked;
+
+    @Setter
     private int stock;
     private Long views;
     @Column(name = "created_at")
